+++ src/blocking_impl.rs
@@ -71,12 +71,6 @@ impl YahooConnector {
         Ok(YOptionResults::scrape(&resp))
     }
 
-    /// Get quote summary for a list of given tickers
-    pub fn get_summary<T: AsRef<str> + Display>(&self, ticker: &[T]) -> Result<YQuoteResponse, YahooError> {
-        let url = format!(YQUOTE_QUERY!(), url = self.quote_url, symbols=ticker_as_list(ticker));
-        YQuoteResponse::from_json(self.send_request(&url)?)
-    }
-
     /// Send request to yahoo! finance server and transform response to JSON value
     fn send_request(&self, url: &str) -> Result<serde_json::Value, YahooError> {
         let resp = self.client.get(url).send()?;
@@ -229,13 +223,5 @@ mod tests {
         assert_eq!(&response.chart.result[0].meta.symbol, "VTSAX");
         assert_eq!(&response.chart.result[0].meta.range, "1mo");
         assert_eq!(&response.chart.result[0].meta.data_granularity, "1d");
-    }
-
-    #[test]
-    fn test_quote_summary() {
-        let ticker = vec!["BTC-USD".to_string(), "BF.B".to_string(), "AAPL".to_string(), "HNL.DE".to_string()];
-        let provider = YahooConnector::new();
-        let response = provider.get_summary(&ticker).unwrap();
-        assert_eq!(response.result.len(), 4);
     }
 }